Make LAZ point counts user-friendly and 64-bit safe
The layer table rendered point counts via Number.toLocaleString(), which
defaults to the 'g' format and showed large counts in scientific notation
(e.g. 1.4e+06). The app also forces the C locale (cwSettings) for stable
serialization, so toLocaleString() won't group digits. Add a page-local
formatPointCount() that inserts comma thousands separators directly, used
by both the wide and narrow delegates.

Also widen cwLazLayer::pointCount() and its Q_PROPERTY from int to qint64
so clouds above ~2.1B points no longer overflow; cwGeometry::vertexCount()
already returns qsizetype, so this just stops the narrowing at the layer
boundary. The QML delegates take pointCount as real (no qint64 keyword in
QML; double is exact well past any realistic count).

Add data-driven formatter tests (0 through 10 billion) and an integration
test asserting the rendered cell text routes through formatPointCount,
guarded by a delegate-materialization wait.

Tests: cavewhere-qml-test tst_GeospatialLayerPage (23 passed),
cavewhere-test [Laz] (319 assertions, 34 cases).

+++ cavewherelib/src/cwLazLayer.h
@@ -56,7 +56,7 @@ class CAVEWHERE_LIB_EXPORT cwLazLayer : public QObject
     Q_PROPERTY(QString sourceCSDisplayName READ sourceCSDisplayName NOTIFY sourceCSChanged)
     Q_PROPERTY(LoadStatus loadStatus READ loadStatus NOTIFY loadStatusChanged)
     Q_PROPERTY(QString errorMessage READ errorMessage NOTIFY errorMessageChanged)
-    Q_PROPERTY(int pointCount READ pointCount NOTIFY pointCountChanged)
+    Q_PROPERTY(qint64 pointCount READ pointCount NOTIFY pointCountChanged)
     Q_PROPERTY(QVector3D bboxMin READ bboxMin NOTIFY bboxChanged)
     Q_PROPERTY(QVector3D bboxMax READ bboxMax NOTIFY bboxChanged)
     Q_PROPERTY(float meanSpacingXY READ meanSpacingXY NOTIFY meanSpacingXYChanged)
@@ -107,7 +107,7 @@ public:
 
     LoadStatus loadStatus() const { return m_loadStatus; }
     QString errorMessage() const { return m_errorMessage; }
-    int pointCount() const { return m_geometry.vertexCount(); }
+    qint64 pointCount() const { return m_geometry.vertexCount(); }
     QVector3D bboxMin() const { return m_bboxMin; }
     QVector3D bboxMax() const { return m_bboxMax; }
     float meanSpacingXY() const { return m_meanSpacingXY; }
